% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cidr.R
\name{cidr}
\alias{cidr}
\alias{conformal_idr}
\title{Conformal IDR}
\usage{
conformal_idr(x, y, x_out, y_out = NULL, online = FALSE, weights = NULL)
}
\arguments{
\item{x}{training covariates}

\item{y}{training observations}

\item{x_out}{new covariates}

\item{y_out}{new observations}

\item{online}{logical specifying whether to sequentially update the training
data set. Default is \code{FALSE}. If \code{FALSE}, the same training data is
used for prediction of all the \code{x_out}.}

\item{weights}{non-negative weights assigned to each covariate and
the new covariate. If omitted, equal weights are used. If \code{online} is \code{FALSE},
\code{weights} is a vector of length \code{length(x) + 1} with the last weight
corresponding to the new covariate. If \code{online} is \code{TRUE}, \code{weights} is a list
of vectors of weights, with lengths \code{length(x) + 1} up to \code{length(x) + length(x_out)}.}
}
\value{
An object of class \code{conformal_fit} containing the conformal IDR fit.
}
\description{
Perform conformal isotonic distributional regression (IDR) to generate a probabilistic
prediction for a real-valued outcome variable. IDR assumes an increasing relationship
between the outcome and the covariate.
}
\note{
The sequential computation of ranks requires that g++ version >= 3.3.1 is
installed. The code was run on a Ubuntu system and adaptations to lines 5
and 6 of online_idr_computation may be necessary on different machines.
}
\references{
\emph{Conformal predictive systems:}

Vovk, V., Gammerman, A. and G. Shafer (2022):
`Algorithmic learning in a random world',
Second Series, Chapter 7.
\doi{10.1007/978-3-031-06649-8}

\emph{Isotonic distributional regression:}

Henzi, A., Ziegel, J. and T. Gneiting (2022):
`Isotonic distributional regression',
\emph{Journal of the Royal Statistical Society: Series B} 83, 963--993.
\doi{10.1111/rssb.12450}

\emph{Conformal IDR:}

Allen, S., Gavrilopolous, G., Henzi, A. and J. Ziegel (2024+):
`Conformal isotonic distributional regression'.
}
\seealso{
\link{cops} \link{lspm} \link{locb}
}
\author{
Sam Allen
}
