% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops.R
\name{cops}
\alias{cops}
\title{Conformal prediction}
\usage{
cops(x, y, x_out, y_out = NULL, method = c("cidr", "lspm", "locb"), ...)
}
\arguments{
\item{x}{training covariates}

\item{y}{training observations}

\item{x_out}{new covariates}

\item{y_out}{new observations}
}
\value{
Conformal prediction object
}
\description{
Generic function for performing conformal prediction of real-valued outcomes.
}
