% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops_eval.R
\name{cops_eval}
\alias{cops_eval}
\alias{pit}
\alias{crps}
\alias{threshcal}
\alias{thickness}
\alias{get_pint}
\alias{coverage}
\alias{int_score}
\alias{width}
\title{Evaluation of Conformal Predictive Systems}
\usage{
pit(fit, y)

crps(fit, y)

threshcal(fit, thresholds, y)

thickness(fit)

get_pint(fit, alpha, crisp)

coverage(fit, y, alpha, average, crisp)

int_score(fit, y, alpha, crisp)

width(fit, alpha, crisp)
}
\arguments{
\item{fit}{object of class \code{"cops"} to be evaluated}

\item{y}{labels that are predicted by \code{fit}. Default is to use \code{fit$y_out}, if available}

\item{thresholds}{thresholds at which to evaluate threshold calibration}

\item{alpha}{level of prediction intervals to be evaluated}

\item{crisp}{logical specifying whether prediction intervals are to be obtained from
the crips CDF (\code{crisp = TRUE}) or from the bounds of the predictive system (\code{crisp = FALSE});
default is \code{crisp = FALSE}.}
}
\value{
\code{pit()} returns a vector of PIT values

\code{crps()} returns a vector of CRPS values

\code{thickness()} returns a vector of thickness values

\code{threshcal()} returns a matrix of forecast exceedance probabilities at the
 thresholds in \code{thresholds}

\code{plot()} returns a base plot object

\code{get_pint()} returns a matrix containing lower and upper bounds of prediction intervals

\code{cov()} returns a vector of coverage indicators

\code{is()} returns a vector of Interval Score values

\code{width()} returns a vector of prediction interval widths
}
\description{
Generic functions for evaluating conformal predictive systems. This includes functions
to calculate the Continuous Ranked Probability Score (CRPS) and thickness of
the predictive systems, obtain Probability Integral Transform (PIT) values, assess
threshold calibration, and to plot the predictive systems.
}
\details{
Details to be added here

\code{plot()} plots a conformal predictive system in \code{fit}, with \code{index}
}
\examples{

n <- 1000
x <- rnorm(n)
y <- rnorm(n, x, exp(x))

N <- 100
x_out <- rnorm(N)
y_out <- rnorm(N, x_out, exp(x_out))

fit <- conformal_idr(x, y, x_out, y_out)

## plot conformal predictive system
plot(fit)
plot(fit, index = 2)
plot(fit, main = "Example Conformal Predictive System", ylim = c(-0.1, 1.1))


## calculate CRPS values
crps(fit) |> mean()

fit2 <- conformal_idr(x, y, x_out) # if y_out not provided, y must be given as an argument in crps()
crps(fit2, y_out) |> mean()


## calculate PIT values
pit(fit) |> mean()
pit(fit2, y_out) |> mean()


## calculate thickness
thickness(fit) |> mean()
thickness(fit2) |> mean()


## threshold calibration
thresholds <- quantile(y, seq(0.1, 0.9, 0.1))
F_t <- threshcal(fit, thresholds)
F_t2 <- threshcal(fit2, thresholds, y_out)

}
\references{
Allen, S., Gavrilopolous, G., Henzi, A. and J. Ziegel (2025):
`In-sample calibration yields conformal calibration guarantees',
\emph{arXiv pre-print} arXiv:2503.03841
\doi{10.48550/arXiv.2503.03841}
}
\seealso{
\code{\link{cops}}, \code{\link{cidr}}, \code{\link{lspm}}, \code{\link{cbin}}
}
\author{
Sam Allen
}
