% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locb.R
\name{locb}
\alias{locb}
\alias{conformal_bin}
\title{Local binning}
\usage{
conformal_bin(x, y, x_out, k = 1)
}
\arguments{
\item{y}{training observations}

\item{k}{number of bins}

\item{X}{training covariates}

\item{X_ts}{test covariates}
}
\value{
Local binning fit based on y and X, applied to X_ts
}
\description{
Computes conformal prediction set using local binning.
}
