% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbin.R
\name{cbin}
\alias{cbin}
\alias{conformal_bin}
\title{Conformal binning}
\usage{
conformal_bin(x, y, x_out, y_out = NULL, online = FALSE, weights = NULL, k = 1)
}
\arguments{
\item{x}{training covariates}

\item{y}{training labels}

\item{x_out}{new covariates}

\item{y_out}{new labels}

\item{online}{logical specifying whether to sequentially update the training
data set. Default is \code{FALSE}.}

\item{weights}{non-negative weights assigned to each training covariate and the
new covariate. If omitted, equal weights are used (the default).}

\item{k}{number of bins}
}
\value{
\code{conformal_bin()} returns a list of objects of class \code{"cops"} containing
the conformal binning fit. See \code{\link{cops}} for details.
}
\description{
Derive conformal predictive systems based on binning. k-means clustering is used
to bin the covariates, and separate predictive systems are then constructed for each bin.
}
\details{
Details to be added here
}
\examples{

n <- 1000
x <- rnorm(n)
y <- rnorm(n, x, exp(x))

N <- 100
x_out <- rnorm(N)
y_out <- rnorm(N, x_out, exp(x_out))
plot(x_out, y_out)

fit <- conformal_bin(x, y, x_out, y_out, k = 10)
fit1 <- conformal_bin(x, y, x_out, y_out)
fit5 <- conformal_bin(x, y, x_out, y_out, k = 5)
fit100 <- conformal_bin(x, y, x_out, y_out, k = 100)

# compare for different k
crps_vec <- c("k = 1" = sapply(fit1, function(x) x$crps) |> mean(),
              "k = 5" = sapply(fit5, function(x) x$crps) |> mean(),
              "k = 10" = sapply(fit, function(x) x$crps) |> mean(),
              "k = 100" = sapply(fit100, function(x) x$crps) |> mean())
print(crps_vec)

}
\references{
\emph{Conformal binning:}

Allen, S., Gavrilopolous, G., Henzi, A. and J. Ziegel (2024+):
`Conformal isotonic distributional regression'.


\emph{Mondrian predictive systems:}

Bostroem, H., Johansson, U. and T. Loefstroem (2021):
`Mondrian conformal predictive systems',
\emph{Proceedings of Machine Learning Research} 152: 1--15.
}
\seealso{
\code{\link{cops}} \code{\link{cidr}} \code{\link{lspm}}
}
\author{
Sam Allen
}
