% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cidr.R
\name{cidr}
\alias{cidr}
\alias{conformal_idr}
\alias{plot.conformal_idrfit}
\title{Online conformal IDR}
\usage{
conformal_idr(x, y, x_out, y_out, online = TRUE, weights = NULL)

\method{plot}{conformal_idrfit}(fit, index = 1)
}
\arguments{
\item{x}{training covariates}

\item{y}{training observations}

\item{x_out}{new covariates}

\item{y_out}{new observations}

\item{online}{whether to sequentially update the training data set. Default
is \code{TRUE}. If \code{FALSE}, the same training data is used for
predicting with all the \code{x_out}.}

\item{weights}{if \code{online} is \code{FALSE}, a vector of the same length
as \code{x} plus one containing non-negative weights, with the last
weight for the new covariate. If \code{online} is \code{TRUE}, a list
of vectors of weights, with lengths \code{length(x) + 1} up to
\code{length(x) + length(x_out)}. If omitted, equal weights are used.}
}
\value{
Returns an object of class \code{conformal_idr}. It contains the jump points
of the conditional CDFs, which are either a matrix with \code{length(x_out)}
columns for \code{online == TRUE}, and a vector otherwise. The values of the
conditional CDFs with the corresponding jump points are returned as matrices,
where \code{cdf_lwr}, \code{cdf_upr}, and \code{cdf_oos} are the CDFs at the
lower bound, upper bound, and interpolation of the two for the given value
of \code{x_out[i]}, and \code{cdf_lcnf}, \code{cdf_ucnf} are the lower and
upper bounds from the conformal IDR.
}
\description{
Computes IDR on a training data set and sequentially adds observations and
updates the fit.
}
\note{
The sequential computation of ranks requires that g++ version >= 3.3.1 is
installed. The code was run on a Ubuntu system and adaptations to lines 5
and 6 of online_idr_computation may be necessary on different machines.
}
