% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cidr.R
\name{cidr}
\alias{cidr}
\alias{conformal_idr}
\title{Conformal IDR}
\usage{
conformal_idr(x, y, x_out, y_out = NULL, online = FALSE, weights = NULL)
}
\arguments{
\item{x}{training covariates}

\item{y}{training labels}

\item{x_out}{new covariates}

\item{y_out}{new labels}

\item{online}{logical specifying whether to sequentially update the training
data set. Default is \code{FALSE}.}

\item{weights}{non-negative weights assigned to each training covariate and the
new covariate. If omitted, equal weights are used (the default).}
}
\value{
\code{conformal_cidr()} returns an object of class \code{"cops"} containing the
conformal IDR fit. If \code{online = TRUE}, then a list of \code{"cops"} objects are returned.
See \code{\link{cops}} for details.
}
\description{
Perform conformal isotonic distributional regression (IDR). Returns a conformal
predictive system for a real-valued outcome variable, under the assumption that
there is an increasing relationship between the outcome and the covariates.
}
\details{
Details to be added here
}
\note{
The sequential computation of ranks requires that g++ version >= 3.3.1 is
installed.
}
\examples{

n <- 1000
x <- rnorm(n)
y <- rnorm(n, x, exp(x))

N <- 100
x_out <- rnorm(N)
y_out <- rnorm(N, x_out, exp(x_out))
plot(x_out, y_out)

fit <- conformal_idr(x, y, x_out, y_out)
plot(fit)

# assign heigher weights to more recent covariates
weights <- c((1:n)/n, rep(1, N))
fit <- conformal_idr(x, y, x_out, y_out, weights = weights)
plot(fit)


}
\references{
\emph{Isotonic distributional regression:}

Henzi, A., Ziegel, J. and T. Gneiting (2022):
`Isotonic distributional regression',
\emph{Journal of the Royal Statistical Society: Series B} 83, 963--993.
\doi{10.1111/rssb.12450}

\emph{Conformal IDR:}

Allen, S., Gavrilopolous, G., Henzi, A. and J. Ziegel (2024+):
`Conformal isotonic distributional regression'.
}
\seealso{
\code{\link{cops}} \code{\link{lspm}} \code{\link{cbin}}
}
\author{
Sam Allen
}
