% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cidr.R
\name{cidr}
\alias{cidr}
\alias{conformal_idr}
\title{Conformal IDR}
\usage{
conformal_idr(
  x,
  y,
  x_out,
  y_out = NULL,
  x_est = NULL,
  y_est = NULL,
  online = FALSE,
  weights = NULL
)
}
\arguments{
\item{x}{training covariates}

\item{y}{training labels}

\item{x_out}{new covariates}

\item{y_out}{new labels}

\item{x_est}{covariates in the estimation set}

\item{y_est}{labels in the estimation set}

\item{online}{logical specifying whether to sequentially update the training
data set. Default is \code{FALSE}.}

\item{weights}{non-negative weights assigned to each training covariate and the
new covariate. If omitted, equal weights are used (the default).}
}
\value{
\code{conformal_cidr()} returns an object of class \code{"cops"} containing the
conformal IDR fit. If \code{online = TRUE}, then a list of \code{"cops"} objects are returned.
See \code{\link{cops}} for details.
}
\description{
Perform conformal isotonic distributional regression (IDR). Returns a conformal
predictive system for a real-valued outcome variable, under the assumption that
there is an increasing relationship between the outcome and the covariates.
}
\details{
Conformal IDR generates conformal predictive systems based on isotonic distributional
regression (IDR). IDR is a non-parametric distributional regression procedure that
assumes an isotonic (increasing) relationship between the covariates and the labels.
Conformal IDR provides a means to generate predictive system bands that are guaranteed
to contain an isotonically-calibrated forecast distribution.

Since conformal IDR requires an ordering of the covariates, it is easiest to implement
when the covariates are univariate real-valued (i.e. \code{x} is a numeric vector).
In the split conformal setting, the estimation data (\code{x_est} and \code{y_est})
can be used to fit a univariate single index model for the label given the covariates.
This model can then be applied to the training covariates (and new covariate) to
obtain univariate index values, and conformal IDR can then be implemented assuming an
isotonic relationship between these index values and the label.

If estimation data (\code{x_est} and \code{y_est}) is provided, and \code{x} is a matrix
or data frame containing multiple covariate variables, then a linear model is fit to the
estimation data, which is used to convert \code{x} and \code{x_out} to univariate index values.
Conformal IDR is performed using these index values.

The linear model is fit using all covariate variables using \code{\link{lm}}.
If a different index model is desired, then this should be fit manually to \code{x_est}
and \code{y_est}, before inputting the resulting index values as \code{x} and
\code{x_out} to \code{conformal_idr()} in the full setting (without \code{x_est} and \code{y_est}).
}
\note{
The sequential computation of ranks requires that g++ version >= 3.3.1 is
installed.
}
\examples{

n <- 1000
x <- rnorm(n)
y <- rnorm(n, x, exp(x))

N <- 100
x_out <- rnorm(N)
y_out <- rnorm(N, x_out, exp(x_out))
plot(x_out, y_out)

fit <- conformal_idr(x, y, x_out, y_out)
plot(fit)

# assign heigher weights to more recent covariates
weights <- c((1:n)/n, rep(1, N))
fit <- conformal_idr(x, y, x_out, y_out, weights = weights)
plot(fit)


}
\references{
\emph{Isotonic distributional regression:}

Henzi, A., Ziegel, J. and T. Gneiting (2022):
`Isotonic distributional regression',
\emph{Journal of the Royal Statistical Society: Series B} 83, 963--993.
\doi{10.1111/rssb.12450}

\emph{Conformal IDR:}

Allen, S., Gavrilopolous, G., Henzi, A. and J. Ziegel (2025):
`In-sample calibration yields conformal calibration guarantees',
\emph{arXiv pre-print} arXiv:2503.03841
\doi{10.48550/arXiv.2503.03841}
}
\seealso{
\code{\link{cops}} \code{\link{lspm}} \code{\link{cbin}}
}
\author{
Sam Allen
}
