% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_meth.R
\name{conformal}
\alias{conformal}
\alias{pit}
\alias{crps}
\alias{threshcal}
\alias{thickness}
\title{Conformal predictive systems}
\usage{
pit(fit, y)

crps(fit, y)

threshcal(fit, y, thresholds)

thickness(fit)
}
\arguments{
\item{x_tr}{vector of covariate values for training}

\item{y_tr}{vector of observations corresponding to `x_tr`}

\item{x_ts}{vector of covariate values for prediction}

\item{online}{logical specifying whether to perform online or offline prediction}

\item{...}{additional arguments to other methods}
}
\value{
An object of type 'conformal_fit'
}
\description{
Generic function for performing conformal prediction of real-valued outcomes.

\code{ConformalIDR} provides methods to obtain conformal predictive systems using
a local binning procedure, least squares predction machine, or conformal isotonic
distributional regression.
}
\details{
Details to be added here
}
\references{
References to be added here
}
\seealso{
\code{\link{cidr}}, \code{\link{lspm}}, \code{\link{locb}}
}
