% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locb.R
\name{locb}
\alias{locb}
\alias{fit_locb}
\title{Local binning}
\usage{
fit_locb(y, X, X_ts, k = 1)
}
\arguments{
\item{y}{training observations}

\item{X}{training covariates}

\item{X_ts}{test covariates}

\item{k}{number of bins}
}
\value{
Local binning fit based on y and X, applied to X_ts
}
\description{
Computes conformal prediction set using local binning.
}
