% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lspm.R
\name{lspm}
\alias{lspm}
\alias{conformal_lspm}
\title{Least squares prediction machine}
\usage{
conformal_lspm(x, y, x_out, y_out = NULL)
}
\arguments{
\item{x}{training covariates}

\item{y}{training observations}

\item{x_out}{new covariates}

\item{y_out}{new observations}
}
\value{
An object of class \code{conformal_fit} containing the LSPM fit.
}
\description{
Performs conformal prediction using the least squares prediction machine (LSPM) algorithm.
}
\references{
\emph{Conformal predictive systems:}

Vovk, V., Gammerman, A. and G. Shafer (2022):
`Algorithmic learning in a random world',
Second Series, Chapter 7.
\doi{10.1007/978-3-031-06649-8}
}
\seealso{
\link{cops} \link{cidr} \link{locb}
}
\author{
Sam Allen
}
