% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lspm.R
\name{lspm}
\alias{lspm}
\alias{fit_lspm}
\title{Least squares prediction machine}
\usage{
fit_lspm(y, X, X_ts)
}
\arguments{
\item{y}{training observations}

\item{X}{training covariates}

\item{X_ts}{test covariates}
}
\value{
LSPM fit based on y and X, applied to X_ts
}
\description{
Computes conformal prediction set using the least squares prediction algorithm.
}
